nearby cells search added

+++ src/fluid.h
@@ -15,6 +15,6 @@ struct pairHash{
     }
 };
 
-void findNeighbors(std::vector<float>& centers, std::unordered_map<std::pair<int, int >, std::vector<float>, pairHash>& grid, int numCircles);
+void findNeighbors(std::vector<float>& centers, std::unordered_map<std::pair<int, int >, std::vector<int>, pairHash>& grid, int numCircles);
 
 #endif
